# benefit-contract/src/main/resources/openapi/benefit-center-v1.yaml
openapi: 3.0.3
info:
  title: Enterprise Benefit Center API
  version: 1.0.0
  description: |
    Company-wide idempotent benefit fulfillment API. Tenant identity is resolved from a verified JWT;
    callers cannot select a tenant or fulfillment channel in the body. V1 package policy is BEST_EFFORT.
servers:
  - url: https://benefit.example.internal
security:
  - bearerAuth: []
...
paths:
  /openapi/v1/award-orders:
    post:
      tags: [Awards]
      operationId: acceptAwardIntent
      summary: Idempotently accept an award intent
      description: Idempotency-Key must equal sourceRequestId in v1. A replay returns the original order.
      security:
        - bearerAuth: [benefit.award.write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AwardIntent' }
      responses:
        '202':
          description: Accepted, including an idempotent replay
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Awards]
      operationId: findAwardOrderBySource
      summary: Recover an order by the producer idempotency identity
      security:
        - bearerAuth: [benefit.award.read]
      parameters:
        - in: query
          name: sourceSystem
          required: true
          schema: { type: string, maxLength: 64 }
        - in: query
          name: sourceRequestId
          required: true
          schema: { type: string, maxLength: 128 }
      responses:
        '200':
          description: Found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AwardOrder' }
        '404': { $ref: '#/components/responses/NotFound' }
  /openapi/v1/award-orders/{awardOrderNo}:
    get:
      tags: [Awards]
      operationId: getAwardOrder
      summary: Query an award order and its item-level outcomes
      security:
        - bearerAuth: [benefit.award.read]
      parameters:
        - in: path
          name: awardOrderNo
          required: true
          schema: { type: string, maxLength: 64 }
      responses:
        '200':
          description: Found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AwardOrder' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /admin/v1/skus/{skuId}:
    put:
      tags: [Administration]
      operationId: upsertBenefitSku
      description: PENDING_APPROVAL templates are immutable. Clients cannot move DRAFT templates to PENDING_APPROVAL or ACTIVE.
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - in: path
          name: skuId
          required: true
          schema: { type: string, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SkuCommand' }
      responses:
        '204': { description: Saved }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/v1/skus/{skuId}:submit-for-approval:
    post:
      tags: [Administration]
      operationId: submitBenefitSkuForApproval
      description: Atomically moves a DRAFT template to PENDING_APPROVAL and enqueues workflow.command.start.v1.
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - in: path
          name: skuId
          required: true
          schema: { type: string, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SkuSubmitCommand' }
      responses:
        '202':
          description: Approval submission accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SkuSubmitAcceptance' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/v1/tenants/{tenantId}:
    put:
      tags: [Administration]
      operationId: upsertTenantPlacement
      description: The path tenant must equal the tenant mapped from the authenticated JWT audience.
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - in: path
          name: tenantId
          required: true
          schema: { type: string, maxLength: 64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TenantCommand' }
      responses:
        '204': { description: Saved }
  /admin/v1/routes/{routeId}:
    put:
      tags: [Administration]
      operationId: upsertChannelRoute
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - in: path
          name: routeId
          required: true
          schema: { type: string, maxLength: 64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteCommand' }
      responses:
        '204': { description: Saved }
  /admin/v1/inventory/adjustments:
    post:
      tags: [Administration]
      operationId: adjustCenterInventory
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InventoryCommand' }
      responses:
        '202': { description: Adjustment appended to the inventory ledger }
...
  /admin/v1/code-assets:
    get:
      tags: [Console]
      operationId: listCodeAssets
      description: Returns metadata only. cipherText is never included; codeHash is truncated to 8 characters.
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - in: query
          name: skuId
          schema: { type: string, maxLength: 128 }
        - $ref: '#/components/parameters/ConsoleLimit'
      responses:
        '200':
          description: Code asset metadata
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CodeAssetView' }
    post:
      tags: [Administration]
      operationId: importEncryptedCodeAsset
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CodeAssetCommand' }
      responses:
        '202': { description: Imported; plaintext codes are not accepted by this contract }
...
components:
  securitySchemes:
    bearerAuth:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.example.internal/oauth/token
          scopes:
            benefit.award.write: Submit award intents
            benefit.award.read: Query award state
            benefit.remediate: Submit and query remediation
            benefit.admin: Manage catalog and owned inventory
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
...
    ConsoleLimit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
...
  responses:
    BadRequest:
      description: Invalid intent or command
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Missing or invalid authentication
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Missing scope, invalid tenant mapping, or cross-tenant envelope
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Idempotency payload conflict or invalid state transition
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found in the authenticated tenant
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    AwardIntent:
      type: object
      additionalProperties: false
      required: [schemaVersion, sourceSystem, sourceRequestId, recipientRef, partialPolicy, items]
      properties:
        schemaVersion: { type: string, enum: ['1.0'] }
        sourceSystem: { type: string, minLength: 1, maxLength: 64 }
        sourceRequestId: { type: string, minLength: 1, maxLength: 128 }
        sourceBusinessNo: { type: string, maxLength: 128, nullable: true }
        recipientRef: { type: string, minLength: 1, maxLength: 256 }
        decision: { $ref: '#/components/schemas/DecisionReference' }
        partialPolicy: { type: string, enum: [BEST_EFFORT] }
        items:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: '#/components/schemas/AwardItemIntent' }
        trace:
          type: object
          additionalProperties: { type: string, maxLength: 256 }
    DecisionReference:
      type: object
      additionalProperties: false
      properties:
        decisionId: { type: string, maxLength: 128 }
        activityId: { type: string, maxLength: 64 }
        activityVersion: { type: integer, format: int32, minimum: 1 }
    AwardItemIntent:
      type: object
      additionalProperties: false
      required: [clientItemId, benefitSkuId, benefitType, quantity]
      properties:
        clientItemId: { type: string, minLength: 1, maxLength: 128 }
        benefitSkuId: { type: string, minLength: 1, maxLength: 128 }
        benefitType: { $ref: '#/components/schemas/BenefitType' }
        amountMinor: { type: integer, format: int64, minimum: 1, nullable: true }
        currency: { type: string, pattern: '^[A-Z]{3}$', nullable: true }
        quantity: { type: integer, format: int64, minimum: 1, maximum: 1 }
        metadata:
          type: object
          additionalProperties: { type: string, maxLength: 256 }
      description: Atomic item (quantity=1). Split multiple units into stable clientItemIds. CASH requires amountMinor/currency; every non-cash type forbids both monetary fields.
    BenefitType:
      type: string
      enum: [CASH, COUPON, SERVICE_VOUCHER, REDEMPTION_CODE, PHYSICAL]
    AcceptResult:
      type: object
      required: [awardOrderNo, status, replay]
      properties:
        awardOrderNo: { type: string }
        status: { $ref: '#/components/schemas/AwardOrderStatus' }
        replay: { type: boolean }
    AwardOrder:
      type: object
      required: [orderNo, sourceSystem, sourceRequestId, recipientRef, status, homeCell, items]
      properties:
        orderNo: { type: string }
        sourceSystem: { type: string }
        sourceRequestId: { type: string }
        sourceBusinessNo: { type: string, nullable: true }
        recipientRef: { type: string }
        status: { $ref: '#/components/schemas/AwardOrderStatus' }
        homeCell: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/AwardItem' }
    AwardOrderStatus:
      type: string
      enum: [ACCEPTED, PROCESSING, SUCCEEDED, PARTIAL_SUCCEEDED, FAILED, REMEDIATING, REVERSING, REVERSED, PARTIALLY_REVERSED]
    AwardItem:
      type: object
      properties:
        itemNo: { type: string }
        clientItemId: { type: string }
        skuId: { type: string }
        benefitType: { $ref: '#/components/schemas/BenefitType' }
        quantity: { type: integer, format: int64 }
        amountMinor: { type: integer, format: int64, nullable: true }
        currency: { type: string, nullable: true }
        status:
          type: string
          enum: [PENDING, RESERVED, DISPATCHING, SUCCEEDED, FAILED_FINAL, UNKNOWN, QUERYING, REISSUING, REVERSING, REVERSED, REVERSAL_FAILED, REVERSAL_UNKNOWN]
        routeId: { type: string, nullable: true }
        failureCode: { type: string, nullable: true }
        latestOperationNo: { type: string, nullable: true }
        latestOperationStatus: { type: string, nullable: true }
        skuVersion: { type: integer, format: int64, minimum: 0 }
        walletEntryId: { type: string, nullable: true }
...
    TenantCommand:
      type: object
      required: [tenantId, homeCell, enabled]
      properties:
        tenantId: { type: string, minLength: 1, maxLength: 64 }
        homeCell: { type: string, minLength: 1, maxLength: 32 }
        enabled: { type: boolean }
        expectedVersion: { type: integer, format: int64, nullable: true }
    SkuCommand:
      type: object
      required: [skuId, benefitType]
      properties:
        skuId: { type: string }
        benefitType: { $ref: '#/components/schemas/BenefitType' }
        faceValueMinor: { type: integer, format: int64, nullable: true }
        currency: { type: string, nullable: true }
        enabled: { type: boolean, nullable: true, deprecated: true, description: Compatibility input; status is authoritative. }
        status: { $ref: '#/components/schemas/SkuTemplateStatus' }
        validityType: { $ref: '#/components/schemas/ValidityType' }
        validFrom: { type: string, format: date-time, nullable: true }
        validTo: { type: string, format: date-time, nullable: true }
        relativeDays: { type: integer, minimum: 1, nullable: true }
        usableWeekdays:
          type: array
          uniqueItems: true
          items: { type: integer, minimum: 1, maximum: 7 }
        dailyQuota: { type: integer, format: int64, minimum: 1, nullable: true }
        userLimitPerDay: { type: integer, format: int64, minimum: 1, nullable: true }
        userLimitTotal: { type: integer, format: int64, minimum: 1, nullable: true }
        equivalentSkuId: { type: string, maxLength: 128, nullable: true }
        expectedVersion: { type: integer, format: int64, nullable: true }
    SkuSubmitCommand:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    SkuSubmitAcceptance:
      type: object
      additionalProperties: false
      required: [skuId, status, version]
      properties:
        skuId: { type: string, maxLength: 128 }
        status: { type: string, enum: [PENDING_APPROVAL] }
        version: { type: integer, format: int64, minimum: 1 }
    SkuTemplateStatus:
      type: string
      enum: [DRAFT, PENDING_APPROVAL, ACTIVE, PAUSED, RETIRED]
    ValidityType:
      type: string
      enum: [ABSOLUTE, RELATIVE]
    RouteCommand:
      type: object
      required: [routeId, skuId, priority, channelCode, ownerType, reserveMode, enabled]
      properties:
        routeId: { type: string }
        skuId: { type: string }
        priority: { type: integer, minimum: 1 }
        channelCode: { type: string }
        ownerType: { type: string, enum: [CENTER_QUOTA, CHANNEL_SHADOW, CENTER_STOCK] }
        fallbackRouteId: { type: string, nullable: true }
        reserveMode: { type: string, enum: [LAZY, EAGER] }
        enabled: { type: boolean }
        configRef: { type: string, nullable: true }
        expectedVersion: { type: integer, format: int64, nullable: true }
    InventoryCommand:
      type: object
      required: [accountId, skuId, ownerType, ownerId, deltaAvailable, requestId]
      properties:
        accountId: { type: string }
        skuId: { type: string }
        ownerType: { type: string, enum: [CENTER_QUOTA, CHANNEL_SHADOW, CENTER_STOCK] }
        ownerId: { type: string }
        deltaAvailable: { type: integer, format: int64 }
        requestId: { type: string }
    CodeAssetCommand:
      type: object
      required: [codeAssetId, skuId, codeHash, cipherText, keyVersion]
      properties:
        codeAssetId: { type: string }
        skuId: { type: string }
        codeHash: { type: string, description: SHA-256 or keyed hash of the code }
        cipherText: { type: string, description: KMS-encrypted code; plaintext is forbidden }
        keyVersion: { type: string }
        expiresAt: { type: string, format: date-time, nullable: true }
...
    CodeAssetView:
      type: object
      required: [codeAssetId, skuId, status]
      properties:
        codeAssetId: { type: string }
        skuId: { type: string }
        status: { type: string }
        codeHashPrefix: { type: string, nullable: true, description: First 8 characters of codeHash. Never the full hash or cipherText. }
        expiresAt: { type: string, format: date-time, nullable: true }
    Error:
      type: object
      description: RFC 9457 Problem Details with a stable business code.
      required: [type, title, status, detail, code, timestamp]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, format: int32 }
        detail: { type: string }
        instance: { type: string, format: uri, nullable: true }
        code:
          type: string
          enum: [INVALID_INTENT, IDEMPOTENCY_PAYLOAD_CONFLICT, SKU_NOT_FOUND, SKU_NOT_ACTIVE, SKU_NOT_DRAFT, SKU_VERSION_CONFLICT, SKU_APPROVAL_LOCKED, SKU_ILLEGAL_TRANSITION, SKU_SUBMIT_IN_FLIGHT, WALLET_ENTRY_NOT_FOUND, WALLET_ILLEGAL_TRANSITION, WALLET_ALREADY_USED, WALLET_VERSION_CONFLICT, WALLET_EXPIRED, USER_LIMIT_EXCEEDED, INVENTORY_INSUFFICIENT, CHANNEL_RETRYABLE, CHANNEL_FINAL_FAILURE, CHANNEL_RESULT_UNKNOWN, REMEDIATION_NOT_ALLOWED, INTERNAL_ERROR]
        message: { type: string, deprecated: true, description: Compatibility alias of detail. }
        timestamp: { type: string, format: date-time }